installer: Skipping Kitty installation if present

+++ install/app-kitty.sh
@@ -1,3 +1,5 @@
+command which -s kitty && echo "Kitty is installed. Skipping." && return
+
 curl -L https://sw.kovidgoyal.net/kitty/installer.sh | sh /dev/stdin launch=n
 
 ln -sf ~/.local/kitty.app/bin/{kitty,kitten} ~/.local/bin/
